Extract io_utils and pwm modules from the monolith (split PR 1/N)
First step of the monolith split. Moves the lowest-risk pure utilities
out of tfbs_footprinter3/tfbs_footprinter3.py into dedicated modules,
keeping the existing import path working via re-exports.

- tfbs_footprinter3/io_utils.py:
    signal_handler, load_json, dump_json, load_msgpack,
    directory_creator, is_online, overlap_range, distance_solve
- tfbs_footprinter3/pwm.py:
    pwm_maker, PWM_scorer
- tfbs_footprinter3/tfbs_footprinter3.py: deletes those definitions,
  re-exports via 'from tfbs_footprinter3.io_utils import ...' so
  existing callers (tests/, hpc/, downstream) are unaffected.

Behavior-neutral: 21 pytest tests still green, smoke import still resolves
main() and re-exported names, ruff clean, no behavior changes. Also drops
three now-unused top-level imports (sys, socket, msgpack) that ruff found
after the move.

Monolith went from 3306 -> 3191 lines. The next extraction passes will
cover ensembl (careful: monkey-patch boundary), alignment, data loaders,
scoring, plotting, and finally main itself into cli.py.
